backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
et_xmlfile==2.0.0
gunicorn==23.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyJWT==2.9.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
